move cost table to cost folder, and move some helpers
