+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallPage/cost/index.js
@@ -1,3 +1,4 @@
+export * from './CostTable';
 export * from './costTypes';
 export * from './costUtils';
 export * from './TraceCostStats';
